When adding folders to response it shouldn’t be recursive.
This was causing performance problems for us because it was loading all the files in a hierarchy.
It isn’t needed as elfinder will make additional requests to load the additional data.

The commands in which it’s used (open, tree) are only supposed to return one level. And the parents command does the walking back up the tree itself.