In the open command allow a custom URL to be used.
In the open command you can supply a URL through which the selected file should be accessed. This is made optional by only adding it if the value is anything other than null.

We need this for Sakai so that we can have people use elfinder to pick a file and then we embed this file in some HTML, but we don’t want the file to be served up all the time through the elfinder connector.