Switch to doing the InputStream reading in Volume.
This means that if you can’t get an OutputStream to write to you can still implement a Volume as all you have todo is read from the InputStream and put is somewhere.

It also reduces code duplication and has try/finally so that we try not to leak streams.