Prevent runaway thread on bad FsService implmenetaion.
We have a custom implementation of FsService and under some circumstances the call to getParent() would continue to return items that never made it back to a root node. This resulted in a runaway thread. We have fixed this issue within our implementation of FsService, but having a safety net to prevent a runaway  thread is good option.